requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
streamlit>=1.28.0
streamlit-chat>=0.1.1
